+++ demos/manual/src/main/java/is/codion/framework/demos/manual/notes/NotesDemo.java
@@ -19,8 +19,6 @@
 package is.codion.framework.demos.manual.notes;
 
 import is.codion.common.db.database.Database;
-import is.codion.common.model.condition.ConditionModel;
-import is.codion.common.model.condition.ConditionModel.Wildcard;
 import is.codion.common.user.User;
 import is.codion.common.version.Version;
 import is.codion.dbms.h2.H2DatabaseFactory;
@@ -182,10 +180,6 @@ public final class NotesDemo {
 		private NoteTableModel(EntityConnectionProvider connectionProvider) {
 			super(new NoteEditModel(connectionProvider));
 			onInsert().set(OnInsert.ADD_TOP_SORTED);
-			// Case-insensitive note search with automatic wildcards
-			ConditionModel<String> noteCondition = queryModel().conditions().get(Note.NOTE);
-			noteCondition.caseSensitive().set(false);
-			noteCondition.wildcard().set(Wildcard.PREFIX_AND_POSTFIX);
 		}
 	}